+++ Cargo.toml
@@ -8,6 +8,7 @@ resolver = "2"
 
 [workspace.dependencies]
 bitflags = "2"
+cursor-icon = "1.1.0"
 serde = { version = "1", features = ["serde_derive"] }
 smol_str = "0.2.0"
 winit-core = { path = "./winit-core", default-features = false, features = ["std"] }
