+++ winit/Cargo.toml
@@ -62,7 +62,7 @@ cfg_aliases = "0.2.0"
 
 [dependencies]
 bitflags.workspace = true
-cursor-icon = "1.1.0"
+cursor-icon.workspace = true
 log = "0.4"
 mint = { version = "0.5.6", optional = true }
 once_cell = "1.12"
