+++ winit/src/window.rs
@@ -16,6 +16,8 @@ pub use crate::icon::{BadIcon, Icon};
 pub use cursor_icon::{CursorIcon, ParseError as CursorIconParseError};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
+#[doc(inline)]
+pub use winit_core::window::*;
 
 /// Represents a window.
 ///
@@ -1625,67 +1627,6 @@ unsafe impl rwh_04::HasRawWindowHandle for Window {
     }
 }
 
-/// The behavior of cursor grabbing.
-///
-/// Use this enum with [`Window::set_cursor_grab`] to grab the cursor.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum CursorGrabMode {
-    /// No grabbing of the cursor is performed.
-    None,
-
-    /// The cursor is confined to the window area.
-    ///
-    /// There's no guarantee that the cursor will be hidden. You should hide it by yourself if you
-    /// want to do so.
-    ///
-    /// ## Platform-specific
-    ///
-    /// - **macOS:** Not implemented. Always returns [`ExternalError::NotSupported`] for now.
-    /// - **iOS / Android / Web / Orbital:** Always returns an [`ExternalError::NotSupported`].
-    Confined,
-
-    /// The cursor is locked inside the window area to the certain position.
-    ///
-    /// There's no guarantee that the cursor will be hidden. You should hide it by yourself if you
-    /// want to do so.
-    ///
-    /// ## Platform-specific
-    ///
-    /// - **X11 / Windows:** Not implemented. Always returns [`ExternalError::NotSupported`] for now.
-    /// - **iOS / Android / Orbital:** Always returns an [`ExternalError::NotSupported`].
-    Locked,
-}
-
-/// Defines the orientation that a window resize will be performed.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum ResizeDirection {
-    East,
-    North,
-    NorthEast,
-    NorthWest,
-    South,
-    SouthEast,
-    SouthWest,
-    West,
-}
-
-impl From<ResizeDirection> for CursorIcon {
-    fn from(direction: ResizeDirection) -> Self {
-        use ResizeDirection::*;
-        match direction {
-            East => CursorIcon::EResize,
-            North => CursorIcon::NResize,
-            NorthEast => CursorIcon::NeResize,
-            NorthWest => CursorIcon::NwResize,
-            South => CursorIcon::SResize,
-            SouthEast => CursorIcon::SeResize,
-            SouthWest => CursorIcon::SwResize,
-            West => CursorIcon::WResize,
-        }
-    }
-}
-
 /// Fullscreen modes.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum Fullscreen {
@@ -1693,98 +1634,6 @@ pub enum Fullscreen {
 
     /// Providing `None` to `Borderless` will fullscreen on the current monitor.
     Borderless(Option<MonitorHandle>),
-}
-
-/// The theme variant to use.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum Theme {
-    /// Use the light variant.
-    Light,
-
-    /// Use the dark variant.
-    Dark,
-}
-
-/// ## Platform-specific
-///
-/// - **X11:** Sets the WM's `XUrgencyHint`. No distinction between [`Critical`] and [`Informational`].
-///
-/// [`Critical`]: Self::Critical
-/// [`Informational`]: Self::Informational
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum UserAttentionType {
-    /// ## Platform-specific
-    ///
-    /// - **macOS:** Bounces the dock icon until the application is in focus.
-    /// - **Windows:** Flashes both the window and the taskbar button until the application is in focus.
-    Critical,
-
-    /// ## Platform-specific
-    ///
-    /// - **macOS:** Bounces the dock icon once.
-    /// - **Windows:** Flashes the taskbar button until the application is in focus.
-    #[default]
-    Informational,
-}
-
-bitflags::bitflags! {
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-    pub struct WindowButtons: u32 {
-        const CLOSE  = 1 << 0;
-        const MINIMIZE  = 1 << 1;
-        const MAXIMIZE  = 1 << 2;
-    }
-}
-
-/// A window level groups windows with respect to their z-position.
-///
-/// The relative ordering between windows in different window levels is fixed.
-/// The z-order of a window within the same window level may change dynamically on user interaction.
-///
-/// ## Platform-specific
-///
-/// - **iOS / Android / Web / Wayland:** Unsupported.
-#[derive(Debug, Default, PartialEq, Eq, Clone, Copy)]
-pub enum WindowLevel {
-    /// The window will always be below normal windows.
-    ///
-    /// This is useful for a widget-based app.
-    AlwaysOnBottom,
-
-    /// The default.
-    #[default]
-    Normal,
-
-    /// The window will always be on top of normal windows.
-    AlwaysOnTop,
-}
-
-/// Generic IME purposes for use in [`Window::set_ime_purpose`].
-///
-/// The purpose may improve UX by optimizing the IME for the specific use case,
-/// if winit can express the purpose to the platform and the platform reacts accordingly.
-///
-/// ## Platform-specific
-///
-/// - **iOS / Android / Web / Windows / X11 / macOS / Orbital:** Unsupported.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-#[non_exhaustive]
-pub enum ImePurpose {
-    /// No special hints for the IME (default).
-    Normal,
-    /// The IME is used for password input.
-    Password,
-    /// The IME is used to input into a terminal.
-    ///
-    /// For example, that could alter OSK on Wayland to show extra buttons.
-    Terminal,
-}
-
-impl Default for ImePurpose {
-    fn default() -> Self {
-        Self::Normal
-    }
 }
 
 /// An opaque token used to activate the [`Window`].
